PolyRegressorのインスタンスをbuild_regressorで呼び出すよう変更
